TCQuery: liquidity formulas

+++ packages/xchainpy_thorchain_query/xchainpy2_thorchain_query/models.py
@@ -136,7 +136,7 @@ class Block(NamedTuple):
 
 class ILProtectionData(NamedTuple):
     il_protection: Decimal
-    total_days: int
+    total_days: float
 
 
 class ConstructMemo(NamedTuple):
